% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{cc}
\alias{cc}
\title{Print column classes}
\usage{
cc(df)
}
\arguments{
\item{df}{A data frame or data table}
}
\description{
Prints the class of each column; just saves a small amount of typing
Works with dataframes and data tables
}
\examples{
abc <- as.data.frame(a = c(1, 2, 3, 4, 5),
                     b = c("a", "b", "c", "d", "e")
                    )
cc(abc)
}
