% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excl.glm.R
\name{excl.glm}
\alias{excl.glm}
\title{Easily exclude columns in glm()}
\usage{
excl.glm(depvar, indata, excl.cols = NULL, ...)
}
\arguments{
\item{indata}{Dataset}

\item{excl.cols}{Vector of column names or vector of column numbers to exclude}

\item{Y}{Dependent variable}
}
\description{
Preserve the " Y ~ . " formula in a glm, but exclude specific named columns
from the "." (i.e., exclude a column or columns from the rest of the
variables in the dataset).
}
\examples{
data(infert)
m1 <- excl.glm("conc", Theoph, excl.cols="Subject")
summary(m1)

m2 <- excl.glm("conc", Theoph, excl.cols=c(1,2))
summary(m2)

m3 <- excl.glm("case", infert, excl.cols="Subject", family=binomial())
summary(m3)
}
