% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_color_hue.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Add together two numbers.}
\usage{
gg_color_hue(n, alpha = 1)
}
\arguments{
\item{n}{Number of colors needed}

\item{alpha}{Alpha value for palette}
}
\value{
A vector with n colors
}
\description{
Add together two numbers.
}
\examples{
gg_color_hue(6)
}
