% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printi.R
\name{printi}
\alias{printi}
\title{Print every n instances of i}
\usage{
printi(i, n = 10, stringname = "Iteration ")
}
\arguments{
\item{i}{The object being iterated}

\item{n}{Print every n iterations}

\item{stringname}{Print stringname before i}
}
\description{
Print every n instances of i
}
\examples{
for(i in 1:100){printi(i, 10)}
}
