% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm0var.R
\name{rm0var}
\alias{rm0var}
\title{Remove 0 variance columns or rows}
\usage{
rm0var(df, dfdim = 2)
}
\arguments{
\item{df}{The input data frame}

\item{dfdim}{2 for columns, 1 for rows}
}
\description{
Remove 0 variance columns from a dataframe
}
\examples{
abd <- as.data.frame(list(a=c(1:10),b=c(21:30),d=rep(5,10)))
rm0var(abd)
}
