% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_get.R
\name{sql_get}
\alias{sql_get}
\title{Get MySQL query results in 1 shot}
\usage{
sql_get(selection = "*", table, database = "user_johnsk26", connect = con)
}
\arguments{
\item{selection}{Columns to grab}

\item{table}{Table to choose}

\item{database}{Database on the MySQL server}

\item{connect}{A previously opened connection, see example}
}
\description{
Get MySQL query results in 1 shot
}
\examples{
\dontrun{
## Establish database connection
con <- dbConnect(MySQL(),
                 user = 'myname',
                 password = 'mypassword',
                 host = 'google.com',
                 dbname='mydbname')

## Will pull the entire table "icd_codes"
sql_get("icd_codes")
}
}
