% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system2.R
\name{system2}
\alias{system2}
\title{Print system commands to KnitR Output
Taken from Stack Overflow}
\usage{
system2(...)
}
\description{
Print system commands to KnitR Output
Taken from Stack Overflow
}
\examples{
\\donotrun{
cmd <- paste0(
plink2,
" --bfile ", mega.bedfile,
" --extract ", var.file,
" --recode A",
" --threads ", nthreads,
" --pca 10",
" --out BioMe_htn_variant_subset"
)
}
}
