% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basedjoy.R
\name{basedjoy}
\alias{basedjoy}
\title{(Partial) Reimplementation of the ggjoy package in base R graphics
For those who are allergic to ggplot}
\usage{
basedjoy(density.var, grouping.var, dataset, shrinkfactor = 2.5,
  fill.col = "grey", add.grid = TRUE, gridcolor = "black",
  xlabtext = "", ylabtext = "", addgroupnames = TRUE,
  title = "Based Joy", ...)
}
\arguments{
\item{density.var}{The variable you want to calculate densities for}

\item{grouping.var}{A grouping variable (not: doesn't work with continuous variable yet)}

\item{dataset}{The input dataset}

\item{shrinkfactor}{Controls space between the distributions}

\item{fill.col}{Fill color for the distributions}

\item{add.grid}{Should we continue the x lines past calculated density}

\item{gridcolor}{Color of gridlines, only if add.grid is TRUE}

\item{xlabtext}{Self explanatory}

\item{ylabtext}{Self explanatory}

\item{addgroupnames}{Add names of the grouping variable to Y axis?}

\item{title}{Set "main=title"}
}
\description{
(Partial) Reimplementation of the ggjoy package in base R graphics
For those who are allergic to ggplot
}
\examples{
Using the iris dataset:
basedjoy(density.var="Sepal.Length", grouping.var="Species", data=iris,
mai=c(1,1,1,1))

}
