% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basedjoy.R
\name{basedjoy}
\alias{basedjoy}
\title{(Partial) Reimplementation of the ggjoy package in base R graphics
For those who are allergic to ggplot}
\usage{
basedjoy(density.var, grouping.var, dataset, shrinkfactor = 2.5,
  fill.col = "grey", add.grid = TRUE, gridcolor = "black",
  xlabtext = "", ylabtext = "", addgroupnames = TRUE, global.lwd = 1.5,
  boxtype = "n", title = "", ...)
}
\arguments{
\item{density.var}{The variable you want to calculate densities for}

\item{grouping.var}{A grouping variable (not: doesn't work with continuous variable yet)}

\item{dataset}{The input dataset, currently must be a dataframe}

\item{shrinkfactor}{Controls space between the distributions}

\item{fill.col}{Fill color for the distributions}

\item{add.grid}{Should we continue the x lines past calculated density}

\item{gridcolor}{Color of gridlines, only if add.grid is TRUE}

\item{xlabtext}{Self explanatory}

\item{ylabtext}{Self explanatory}

\item{addgroupnames}{Add names of the grouping variable to Y axis?}

\item{global.lwd}{line width of the density plot and gridlines (if applicable)}

\item{boxtype}{Box drawn around plot, options in c("o","7","l","c","u") for lines or "n" for no lines (default)}

\item{title}{Set "main=title"}
}
\description{
(Partial) Reimplementation of the ggjoy package in base R graphics
For those who are allergic to ggplot
}
\examples{
Using the iris dataset:
basedjoy(density.var="Sepal.Length", grouping.var="Species", data=iris,
mai=c(1,1,1,1))

Using an example dataset from ggjoy
library(ggjoy)
lincoln_df <- as.data.frame(lincoln_weather)
lincoln_df$mean.temp <- as.double(lincoln_df$`Mean Temperature [F]`)
lincoln_df$mnth <- as.factor(lincoln_df$Month)

basedjoy("mean.temp", "mnth", lincoln_df, shrinkfactor=50,
        title="Monthly Temperature in Lincoln", mai=c(0.5,1.25,1,0.5))
}
