% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excl.glm.R
\name{excl.glm}
\alias{excl.glm}
\title{Easily exclude name columns in glm()}
\usage{
excl.glm(depvar, indata, excl.cols = NULL, ...)
}
\arguments{
\item{indata}{The dataset}

\item{excl.cols}{Vector of column names or column numbers to include}

\item{The}{Y variable}
}
\description{
Preserve the " Y ~ . " formula in a glm, but exclude specific named columns
from the "." (i.e., exclude a column or columns from the rest of the
variables in the dataset).
}
\examples{

m1 <- excl.glm("conc", Theoph, excl.cols="Subject")
summary(m1)

m1 <- excl.glm("conc", Theoph, excl.cols=c(1,2))
summary(m1)
}
